TPE-344 - refactor quark commands

+++ ProtonMail/ProtonMailUITests/Utils/DateUtils.swift
@@ -9,7 +9,7 @@
 import Foundation
 
 extension Date {
- var millisecondsSince1970:Int64 {
-        return Int64((self.timeIntervalSince1970 * 1000.0).rounded())
+ var millisecondsSince1970:Int {
+        return Int((self.timeIntervalSince1970 * 1000.0).rounded())
     }
 }
